fix: prepare for release v1.1.0

+++ src/commands/kill.ts
@@ -2,6 +2,7 @@ import { Command, Option } from 'clipanion'
 
 import { LeonInstance } from '../services/LeonInstance.js'
 import { log } from '../services/Log.js'
+import { prompt } from '../services/Prompt.js'
 
 export class KillCommand extends Command {
   static paths = [['kill']]
@@ -14,11 +15,19 @@ export class KillCommand extends Command {
     description: 'Name of the Leon instance.'
   })
 
+  public yes = Option.Boolean('--yes', {
+    description: 'Skip all questions with a "yes" answer.'
+  })
+
   async execute(): Promise<number> {
     try {
+      const { yes = false } = this
       const leonInstance = LeonInstance.get(this.name)
-      await leonInstance.kill()
-      console.log(`Leon instance "${leonInstance.name}" killed.`)
+      console.log(`You are about to kill Leon instance "${leonInstance.name}".`)
+      if (yes || (await prompt.shouldExecute('Are you sure?'))) {
+        await leonInstance.kill()
+        console.log(`Leon instance "${leonInstance.name}" killed.`)
+      }
       return 0
     } catch (error) {
       log.error({
